Fix mobile safari design.
The design was breaking because of a Safari issue with width:100%;
The elements on the page were made so that when we have a `float:left` or a `display:flex` with the widths of all elements making up to 100%. This was causing an issue on Safari - the last element was shown on the next line rather than the same.
All width values were decreased by ~1% of their value to make the total decrease to 99% the width.